tslib: Fix support for 2.6.30 through 2.6.35 Linux kernels
The ABS_MT_SLOT define is not present in kernels before 2.6.36, but
the ABS_MT_SLOT value is used in the library and test applications.
In the library there is a defining of ABS_MT_SLOT for kernels before
2.6.30, but this skips 2.6.30 through 2.6.35.  The tests do not
attempt to define ABS_MT_SLOT so they will fail to compile on any
kernel before 2.6.36.

Adding a define for ABS_MT_SLOT if it is not defined in the tslib.h
file means anything using the tslib library, including the test apps,
will compile without issue on older kernels.